% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_SA2}
\alias{cpp_SA2}
\title{Stochastic approximator}
\usage{
cpp_SA2(
  Y,
  X,
  LINK,
  DICT1,
  DICT2,
  START,
  STEP0,
  STEP1 = 1,
  STEP2 = 1e-20,
  STEP3 = 0.75,
  SCHEDULE = 2L,
  UPDATE = 0L,
  SWITCH = 0L,
  AD1 = 0.9,
  AD2 = 0.999,
  PAIRS_PER_ITERATION = 8L,
  BURNE = 2L,
  MAXE = 4L,
  ISH = TRUE,
  UPE = 100000L,
  SEED = 123L,
  VERBOSE = 1L,
  NCAT = 2L
)
}
\description{
Stochastic approximator
}
