% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullPL.R
\name{fullPL}
\alias{fullPL}
\title{Compute full Pairwise likelihood}
\usage{
fullPL(
  Y,
  X,
  F1,
  F2,
  MODEL = c("probit", "logit", "ordprobit"),
  START,
  VERBOSE = 0
)
}
\arguments{
\item{Y}{Response vector}

\item{X}{Design matrix}

\item{F1}{Row factor vector}

\item{F2}{Column factor vector}

\item{MODEL}{Choose among "probit", "logit" and "ordprobit"}

\item{START}{Starting vector for model parameters}
}
\value{
fitted model
}
\description{
Compute full Pairwise likelihood
}
