% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_list_ind_dict}
\alias{get_list_ind_dict}
\title{Creat factor dictionaries}
\usage{
get_list_ind_dict(F1, F2)
}
\arguments{
\item{F1}{Row factor}

\item{F2}{Column factor}
}
\value{
List with one dictionary per factor
}
\description{
Creat factor dictionaries
}
