% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{marg2cond}
\alias{marg2cond}
\title{From marginal to conditional parameterisation}
\usage{
marg2cond(PAR)
}
\arguments{
\item{PAR}{Marginal parameter vector}
}
\description{
From marginal to conditional parameterisation
}
