% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_sa_traj}
\alias{plot_sa_traj}
\title{Plot stochastic parameters trajectories}
\usage{
plot_sa_traj(FIT, MARG = FALSE)
}
\arguments{
\item{FIT}{Fitted object from \link{saPL}}

\item{MARG}{TRUE for marginal parameterisation. FALSE for conditional}
}
\description{
Plot stochastic parameters trajectories
}
