% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saPL.R
\name{saPL}
\alias{saPL}
\title{Fit via stochastic pairwise likelihood}
\usage{
saPL(
  Y,
  X,
  F1,
  F2,
  MODEL = c("probit", "logit", "ordprobit"),
  START,
  CONTROL = list(),
  VERBOSE = 0
)
}
\arguments{
\item{Y}{Response vector}

\item{X}{Design matrix}

\item{F1}{Vector with row memberships}

\item{F2}{Vector with column memberships}

\item{MODEL}{Choose among "probit", "logit" and "ordprobit"}

\item{START}{Starting vector for model parameters}

\item{CONTROL}{control list to be passed to the optimiser}

\item{VERBOSE}{Verbose output}
}
\description{
Fit via stochastic pairwise likelihood
}
